+++ test/fixture_expectation_test.rb
@@ -85,59 +85,78 @@ class FixtureExpectationTest < Minitest::Test
 
   private
 
-      def fixture_property(element, property)
-        return element.class.name if property == "class"
+  def fixture_property(element, property)
+    return element.class.name if property == "class"
 
-        if element.respond_to?(property)
-          return dump_fixture_value(element.public_send(property))
-        end
+    unless element.respond_to?(property)
+      flunk(
+        "#{element.class} has no property #{property.inspect}"
+      )
+    end
 
-        result = element.interpret
+    dump_fixture_value(element.public_send(property))
+  end
 
-        return dump_fixture_value(result.public_send(property)) if result.respond_to?(property)
+  def dump_fixture_value(value)
+    return dump_array(value) if value.is_a?(Array)
 
-        flunk(
-          "Neither #{element.class} nor " \
-          "#{result.class} has property #{property.inspect}"
-        )
-      end
+    send(dumper_for(value), value)
+  end
 
-      def dump_fixture_value(value)
-        case value
-        when Array
-          value.map { |item| dump_fixture_value(item) }
-        when MifParser::Cell
-          dumped = {
-            "class" => value.class.name,
-            "text" => value.text
-          }
-          dumped["elements"] = dump_fixture_value(value.elements)
-          dumped
-        when MifParser::Paragraph
-          dumped = {
-            "class" => value.class.name,
-            "raw_text" => value.raw_text
-          }
-          dumped["tag"] = value.tag unless value.tag.nil?
-          dumped
-        when MifParser::List
-          dumped = {
-            "class" => value.class.name,
-            "raw_text" => value.raw_text,
-            "list_type" => value.list_type,
-            "list_marker" => value.list_marker
-          }
-          dumped["tag"] = value.tag unless value.tag.nil?
-          dumped
-        when MifParser::Format
-          value.to_h
-        when MifParser::TextRun
-          {
-            "text" => value.text,
-            "format" => dump_fixture_value(value.format)
-          }
-        else
-          value
-        end
-      end
+  def dumper_for(value)
+    {
+      MifParser::Cell => :dump_cell,
+      MifParser::Paragraph => :dump_paragraph,
+      MifParser::List => :dump_list,
+      MifParser::Format => :dump_format,
+      MifParser::TextRun => :dump_text_run
+    }.fetch(value.class, :identity_dump)
+  end
+
+  def dump_array(values)
+    values.map { |item| dump_fixture_value(item) }
+  end
+
+  def dump_format(format)
+    format.to_h
+  end
+
+  def identity_dump(value)
+    value
+  end
+
+  def dump_cell(cell)
+    {
+      "class" => cell.class.name,
+      "text" => cell.text,
+      "elements" => dump_fixture_value(cell.elements)
+    }
+  end
+
+  def dump_paragraph(paragraph)
+    dump_tagged_text(paragraph)
+  end
+
+  def dump_list(list)
+    dump_tagged_text(list).merge(
+      "list_type" => list.list_type,
+      "list_marker" => list.list_marker
+    )
+  end
+
+  def dump_tagged_text(element)
+    dumped = {
+      "class" => element.class.name,
+      "raw_text" => element.raw_text
+    }
+    dumped["tag"] = element.tag unless element.tag.nil?
+    dumped
+  end
+
+  def dump_text_run(run)
+    {
+      "text" => run.text,
+      "format" => dump_fixture_value(run.format)
+    }
+  end
 end